Move other failing tests to the correct place
